+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `CodeGenerator`, `Validation`, templates: optimised decimal comparisons
 - `CodeGenerator`, `Validation`, templates: optimised integer and long range checks
 - `CodeGenerator`: bug fix - no longer outputs checks for strings with format causing them to be other class
+- templates: use `open class` for empty inner class
 
 ## [0.28] - 2021-03-23
 ### Changed